use Hyrax docker base setup
Hyrax now provides a minimal base docker image, intended to be reused and
extended by applications. this is a proposal to replace the existing Dockerfile
with one based on that image.

this would switch the application server to `puma`. our thinking is that real
containerized deployments (e.g. via the Helm chart) are likely to be behind a
load balancer and an `nginx` (an ingress controller).

the image is kept minimal with the idea that a smaller set of system
dependencies means a lower footprint for security risks, besides the bandwidth
and storage convenince. Hyrax's setup definitely has space to improve here, but
the idea is greater consolidation will help the community maintain something
really solid.

there's still a little work to do here:

  - remove dependency on `zk` (and therefore java runtime).
  - refine & use Hyrax's `hyrax-worker` image setup for `sidekiq` workers (using
    a separate lets us consider optimizations like `MALLOC_ARENA_MAX=2`).
  - systematically reconcile environment variables (the job here is a bit
    patchwork).

if this is looking like a good direction, I can work on getting this working
with the upstream helm chart as well.