build, tag, and push a proper image for deployment
the hyku dockerfile changed to have multiple targets, but we were building and
pushing the terminal one, which was a sidekiq worker. this left us with no web
service.

switch the main build stage to build the webapp image. also add a stage to build
and push the worker, so it can also be deployed.