Adding logic to extract text from plain text files
This is a first pass of logic that passes unit tests.  I am putting this
up to close out my work week.

Prior to this commit, we were not handling uploaded plain text files.
The impact was that we were not adding the content of the text file to
the field we were using full text search result.

With this commit, we insert a derivative service class for handling text
files.  This service class short-cirtcuits the logic that spans Hyrax
and Hydra::Derivatives, which relied on Solr's text extraction service.

Acceptance Criteria

- [ ] In the application, attach a plain text file (I recommend on that
ends in `.txt`).  That plain text file should have some unique text.
Then after the file has been uploaded, search for the unique text.

Note: As <2022-10-14 Fri 17:47> I have not tested locally but wanted to
get this up for review.  If someone has time to pull this down and test
locally, that would be wonderful.  If not, I will pick this up later on
and amend the commit message.

Closes #147