Removing duplicate methods in processor chain
Prior to this commit, when we searched for indexed text, we would not
find the works for those files sets that had the extracted text.

With this commit, when we search for indexed text, we will find the
works for the associated files.

What was happening is that the `:add_advanced_parse_q_to_solr` processor
chain method was obliterating the join logic.

Yes there's a duplicate for add_access_controls_to_solr_params; but that
does not appear to be causing a problem like the duplication and order
of the now removed additional `:add_advanced_parse_q_to_solr`,
`:add_advanced_search_to_solr` filters.  Those existed in their current
position and at the end of the array.