Adding job to find non-FTS-ed text files
Prior to this commit, we were full text indexing all plain text files
that were newly uploaded.  However, that was a new feature, and did not
account for existing plain text files that had been uploaded.

With this commit, we're introducing a maintenance task to find all of
the plain text FileSets and add their associated text file into the
file_set's extracted text.

In addition, we're extracting common logic used for both the derivative
service as well as the maintenance task; thus allowing for shared logic
and to provide some connectivity between this commit and the !199 merge
request.

*Testing Instructions*

This one requires that you have a plain text file that was uploaded
prior to the !199 change (e.g. the file has not been indexed for full
text search).

This is likely files that are on a staging environment.  As this is a
maintenance task that we must hand run (e.g. it's a rake task) we'll
need to coordinate how to do that.

However, in making this a job, we'll have insight into the job failures
via the /sidekiq dashboard.

*Notes*

I have tested the inner logic of the job in the staging environment,
which is how I uncovered the `Encoding::UndefinedConversionError` error.

Closes #147.