Enabling the CSV Parser
Prior to this commit, we only enabled the
`Bulkrax::OaiAdventistQdcParser`.

With this commit we're also enabling `Bulkrax::CsvParser`.  We are also
now sharing the mappings for those two parsers.  This change **does not**
address the accuracy of the mappers, only that we have enabled the
mappers.

The idea is that with the CSV and OAI QDC parser using the same
mappings, we can test the CSV parser first (e.g. work out the kinks).
Then test the OAI QDC parser against that resolved CSV parser.  Why?  My
assumption is that testing a CSV is easier than testing an OAI QDC end
point.

Works towards #153.  Relates to #182 and #152.

**Testing Instructions**

In the application's dashboard, go to the importers.  Create, but don't
save, a new importer.  In the drop-down for parsers, you should see two
options:

- OAI - Adventist Digital Library
- CSV - Comma Separated Values