Adding logger to IndexPlainTextFilesJob::One
Prior to this commit, we handled exceptions by raising them and
retrying.  However, in the case of this job, retries are highly unlikely
to resolve the issue.  But they might, which is why there's an altered
retry approach.

With this commit, we're logging to a custom file.  After the job is
complete, we can review (and download) the file to review any errors.