Restoring SETTINGS__SOLR__URL
Prior to this commit, we did not have this property.  It was removed via
061a529f.

However our `ops/env.conf` file appears to reference this value:

```shell
❯ rg SETTINGS__SOLR__URL
ops/env.conf
23:env SETTINGS__SOLR__URL;
```

With this commit, I'm restoring that setting.  I'm unclear if this will
fix the [issue that I encountered and reported in this Slack thread][1].

[1]:https://assaydepot.slack.com/archives/G0311DN2YCA/p1667582787506699