Adding abstract to image and generic work
This commit introduces two primary changes:

1.  A decorator to amend how we handle CSV headers (see
    samvera-labs/bulkrax#669)
2.  Adding an abstract property to GenericWork and Image
    models

The reason these are conflated is that I was testing the Abstract
property by way of our CSV importer.  And in our parser the abstract
comes from the `description.abstract` column.  I tested the importer and
the field and value did not render on the show page.  I edited the work
in the UI, and the HTML form included that abstract field.  I added an
abstract, submitted the HTML form, and the abstract was then present on
the show page.

Relates to:
- #104
- samvera-labs/bulkrax#669
- samvera-labs/bulkrax#670