Adding pagination to IndexPlainTextFilesJob
Prior to this commit, we were likely not leveraging paginated results to
the best of our ability.  How this manifested was that we would see a
spike in memory and the job would trigger an out of memory restart once
we had processed around 400 works.

With this commit, we have a narrow pagination window; which will
hopefully ease the memory burden.