Step 1 of migrating Slug Predicate
Prior to this commit, we have the slug's predicate writing to
`DC.alternate`.  This is a mismapping, and one that will cause issues
with the client's desired `alternate_title` property and the
upgradability of the application (Hyrax 3.x introduces `alternate_title`
that maps to the `DC.predicate`)

With this commit, we create a "shadow" attribute that updates in
the same manner as the `slug`.  This prepares the way for a "schema"
migration by ensuring that the new property (with the correct predicate)
is being written to Fedora.

For reference on the renaming pattern:

- http://www.agiledata.org/essays/renameColumn.html
- https://crucialfelix.github.io/posts/renaming-database-columns-in-production/

Related to #77