Removing Related URLs override that rendered downloaded files
Prior to this commit, the `related_url` property was modified for
showing records.  This modification meant that any values actually set
in the `related_url` would not render (but were searchable).  To set
those, you would need to do so via the UI.

Prior to and after this commit, Bulkrax's import mapping mapped
`remote_files` from the `related_url`.  We use the CSV column of
`related_url` (or XML node in the OAI ADL) to indicate files to download
and attach (as FileSets).  The ingest process does not store these
values in the `related_url` attribute of the Fedora document.

With this commit, we're preserving the import behavior (e.g. mapping
from `related_url` to `remote_files`) and exposing the values for
`related_url` that might have been entered via the UI.

_Note:_ This change highlights that we may have a misunderstanding in
the mappings or could rename the import from field or accept what we
have as an implementation detail/foible.

_Note:_ I normally don't like commenting out a method; however it is
helpful to include that commented out method as a connection to the
potential mapping disconnect.

Relates to #87, which includes an even deeper dive into the issue.