Ensuring AARK renders for Generic Work and Image
Prior to this commit the Image and Generic Work did not expose the AARK
to the show page.  They did, however, expose the AARK for editing.

With this commit, the Image and Generic Work now expose the AARK field
on the show page and leverage the dog biscuits specified translation.

Discovered while attempting to locally QA #111.  There is, however, a
disconnect as the Image and Generic Work use the default SimpleForm i18n
of `aark_id` and the others use the Dog Biscuits translation.  To
normalize this, I added the `label: t('dog_biscuits.fields.aark_id')`
option for rendering the `aark_id`.