Fixing reindexing logic to reflect search is Hash
Prior to this commit, I had assumed that the `FileSet.search_in_batches`
yielded an array of Fedora objects; but I was mistaken.  It was yielding
an array of Hashes.  Which created issues.

After this commit, I'm operating on the hashes and moving a method down
into the worker to account for the fact that the Hash representation of
the FileSet does not have the `extracted_text` in it's data.