Adding DailySolrContentCheckService.call function
With this commit, I'm adding a simple query to write to a buffer tab
separated values.  The columns of this tab separation are:

- cname :: The account's cname (more useful than ID)
- aark_id_tesim :: The record's associated aark_id
- id :: The record's ID; not necessarily the Fedora ID
- timestamp :: when was this item indexed
- fedora_id_ssi :: The record's Fedora ID

Because of SlugBug's implementation, the ID of the SOLR document need
not be the Fedora ID but could be the Slug.

I made a few assumptions that are parameterized:

1. We'll be writing the output to a buffer (e.g. STDOUT)
2. We are only interested in objects that have an AARK_ID

This does not address how we call it nor what we do with the buffer.
That is ambiguous in the current ticket.

Relates to #37