Adding ordering to :creator (labeled as Author)
Prior to this commit, the order that we specified the `:creator` values
were not necessarily remembered when we would round-trip data from Ruby
to Fedora and back to Ruby.

With this commit, we introduce the OrderAlready gem for encoding the
values for an attribute with a sort order.  And then declare the
`:creator` property to be ordered.

This will require reingesting any records that already exist; or
accepting the order that the application chooses when we round trip from
Fedora to Ruby and back to Fedora.

Related to:

- https://github.com/scientist-softserv/adventist-dl/issues/84
- https://github.com/scientist-softserv/adventist-dl/issues/55