Ensuring we have "all_fields" as part of all fields
**NOTE**: I would like to understand more of what lead to us comment out
code; I know I was prodding to remove any DogBiscuits declarations as
they cloud the implementation.

Prior to this commit, the "all_names" local variable was an empty
string.  This emptiness was the result of @1ceee50 commenting out the
following in `app/controllers/catalog_controller.rb`:

```ruby
show_props = DogBiscuits.config.all_properties
add_show_field config, show_props
```

The impact was that this removed almost all fields from the fields in
which we searched for values.

With this commit, instead of uncommenting out the code, which may be the
more correct pathway, I want to mimic what DogBiscuits was doing with
those properties.  In addition, I'm allowing for the potential to
declare additional show_fields that are not part of DogBiscuits; at
present there are none..

This relates to and would eventually close:

- https://github.com/scientist-softserv/adventist-dl/issues/43