Ensuring metadata_prefix option is correctly passed
Prior to this commit, the `metadata_prefix` passed on `OAI::Client.new`
was discarded.  In the current code, that option is not processed.  It
is instead a parameter to pass to the OAI verb
action (e.g. `ListIdentifiers` and `GetRecords`).

With this commit, we ensure that we pass the `metadata_prefix` to the specific verb.

Related to:

- https://github.com/samvera-labs/bulkrax/issues/676