Quasi-backporting of native SOLR Graph nesting
Prior to this commit, we had disabled the logic for indexing the parent
child relationships of objects.  This results in broken behavior
regarding nesting collections and works.  Were we to enable the
"classic" strategy for indexing, we created significant (and known)
performance issues.

With this commit, we are taking code that was backported to the 3.x
branch of Hyrax via [samvera/hyrax#5858][5858] and
[samvera/hyrax#5916][5916] and projecting that onto the 2.9.x
implementation.

This projection means that we are not directly backporting but instead
structuring the code such that if you include it, it assumes you using
SOLR's native graph indexing.  The aforementioned PRs put those changes
behind a configuration variable; which is not something that makes sense
given this "hack."

This PR aims to resolve three issues, which by their nature, are
inter-related.  The issues are:

- [x] Adding works to a collection [#59][59]
- [x] Backporting graph indexing [#68][68]
- [x] Adding a collection to another collection [#70][70]

[59]: https://github.com/scientist-softserv/adventist-dl/issues/59
[68]: https://github.com/scientist-softserv/adventist-dl/issues/68
[70]: https://github.com/scientist-softserv/adventist-dl/issues/70
[5858]: https://github.com/samvera/hyrax/pull/5858/
[5916]: https://github.com/samvera/hyrax/pull/5916/