Updating Bulkrax to handle invisible BOM charactesr
The [Bite Order Mark (BOM)][1] is an invisible character that
interfers with field mapping.  In this case we had a CSV that had the
column named "file" and our mapping had a field named "file".

On a visual examination, this should've matched.  However the CSV had a
BOM character at the beginning of "file" column.  Thus the string
comparison failed because of the invisible character.

The manifestation of which was failure to upload and attach files.

Related to:

- samvera-labs/bulkrax#689
- scientist-softserv/adventist-dl#179

[1]: https://en.wikipedia.org/wiki/Byte_order_mark