Updating OAI parsing for Adventist based on assumptions
Prior to this commit, the logic for the Adventist OAI parsing was a
hybrid of CSV and OAI parsing logic.  The parser logic had faulty
assumptions.

In addition, the `Bulkrax::OaiAdventistQdcEntry` introduced logic that
created collections based on the OAI set.  However, based on the OAI
implementation we should only be creating collections when the set has
an AARK ID.  In the prior implementation we were creating a collection
for with the effective title of `adl:periodical`; and that is not a
desired collection.

With this commit, the core logic better aligns with the outlined
assumptions in the code comments and the OAI feed implementation.  I
have made efforts to document intention and how it relates to the
foibles of Adventist's OAI implementation.

I have taken efforts to properly address the parser's counts/totals but
these things are quite challening given their upstream Bulkrax
implementation.

Related to:

- https://github.com/scientist-softserv/adventist-dl/issues/207
- https://github.com/scientist-softserv/adventist-dl/issues/197
- https://github.com/scientist-softserv/adventist-dl/issues/198
- https://github.com/scientist-softserv/adventist-dl/issues/199
- https://github.com/scientist-softserv/adventist-dl/issues/200