Introduce memoization and better default
Prior to this commit, we were looping through records and calling
model_field_mappings; each time we regenerated the array.  In addition
the `model_mappings ||= ['model']` would never be set because before that we
guaranteed that `model_mappings` was at least `[]`

With this commit, we both memoize the results and set the initial
intended default value.