Adding spring to provide option for faster tests
To leverage spring, you will need to:

- Start the containers (e.g. `sc up`)
- Run the specs using spring (e.g. `spring rspec path/to/my_spec.rb`)

Spring boots up a Rails server and re-uses that between invocations.
This means that instead of spending about 1 minute booting the
application to run a 1 second test, you can re-use that app (spending
about 8 seconds or so connecting to it) then run that 1 second test.

But beware of spring, as it can create odd behavior in the application.
As a matter of practice when you are done testing, run `spring stop` to
terminate the spring booted applications.

Related to:

- https://github.com/scientist-softserv/britishlibrary/pull/252