Bumping Fedora memory allocation
Before this commit, we were seeing `LDP::NotFound` errors which were
part of the failure for this [show entry][1].  The resource that we
didn't find was the tenant's Fedora ID.  Which we know exists.

Note: This is something we encountered in Loiusville ([see Slack
thread][2]).  It has not been resolved in Louisville.

With this commit, we're taking an educated guess as to what will be
adequate for fcrepo.  Note, we're keeping production and staging
synchronized so that we have comparable environments.

References:

- https://github.com/scientist-softserv/dev-ops/issues/551
- https://github.com/scientist-softserv/adventist-dl/issues/240

[1]: https://adl.s2.adventistdigitallibrary.org/importers/36/entries/186125?locale=en

[2]: https://assaydepot.slack.com/archives/G0313NK5NMA/p1664387275241309