Adding spec to verify load order of services
Prior to this commit, I had manually checked but not programmatically
verified the order of the services.

We need to verify this because the first service that says it's valid is
the one the runs (and the others do not run).

Why have the `Adventist::TextFileTextExtractionService` one first?
Because that is the faster one to validate.

Related to:

- https://github.com/scientist-softserv/adventist-dl/issues/117