Skipping `identifier` element in OAI metadata XML element
Prior to this commit we had two "identifier" elements.  In our baseline
`Bulkrax::OaiEntry` implementation, we assign the `header > identifier`
to the `identifier` property (as defined by the bulkrax configuration
for the `source_identifier: true` property.  We also assigned the
`metadata > identifier` element to the `identifier` property based on
the `from: ['identifier']` mapping.

The end result was that our `identifier` property got the following
values: `["13129009", ".b31290097"]`.  Which might cause problems given
the expectations around the [`source_identifier` in Bulkrax][1].

With this commit, we short-circuit the `field_to` logic to return an
empty array.  Thus ignoring the `metadata > identifier` value.  The
`field_to` logic is not called for the `header > identifier` but only
for the `metadata` elements.

[1]: https://github.com/samvera-labs/bulkrax/wiki/CSV-Importer#source-identifier

Below is the XML example to help clarify the above prior and current
state.

```xml
<OAI-PMH xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://www.openarchives.org/OAI/2.0/" xsi:schemaLocation="http://www.openarchives.org/OAI/2.0/ http://www.openarchives.org/OAI/2.0/OAI-PMH.xsd">
  <responseDate>2023-02-15T16:23:07Z</responseDate>
  <request verb="GetRecord">
https://oai.adventistdigitallibrary.org/OAI-script?verb=GetRecord&metadataPrefix=oai_adl&identifier=13129009
  </request>
  <GetRecord>
    <record>
    <header>
    <identifier>13129009</identifier>
    </header>
    <metadata>
    <oai_adl>
    <aark_id>13129009</aark_id>
    <identifier>.b31290097</identifier>
    </oai_adl>
    </metdata>
    </record>
  </GetRecord>
</OAI-PMH>
```