Ensuring that collections can add to collections
Prior to this commit, we were taking the `part_of` property (and any
member of collection attributes) and moving them into the `collection`
property.  The `collection` property was then ignored in later parsing.

The intention was to put the collection identifiers into the
parsed_metadata field that would represent the member of collection
property.  Which, by default in Bulkrax is `parents`.

With this commit, we are mapping the collections into the `parents`
parsed metadata, which in turn will schedule the
`Bulkrax::CreateRelationshipsJob` and thus establish the relationship.

In reviewing #203, we are unclear, but have conjecture, about why there
are additional relationships.  This addresses adding the missing
`part_of` relationship that we were expecting to be present.

Related to:

- https://github.com/scientist-softserv/adventist-dl/issues/203