Fix CSV metadata
In `bulkrax.rb` had `part` instead of `part_of` so the fields were not
mapping correctly, a test was added to the CsvEntrySpec.  Curious, since
in the OAI parser there is a `part` field as well as a `part_of` field.
When checking on the model, I didn't see a `part` property and only a
`part_of` property.

The `rights_statement` was not being rendered properly on the view
template and that has been corrected.