Configuring to use tesseract best training data
Prior to this commit, we're using the default configuration for
Tesseract.

In discussion with the client, we provided examples that used the better
training data (as per the Tesseract documentation).  In comparing
different setups they chose the configiruation to use the `-l eng_best`
and the `--psm 1` configuration.

After running `docker compose build`, when we shell into the docker
container and run =tesseract --list-langs= we see the following:

```bash
bash-5.1$ tesseract --list-langs
List of available languages (4):
eng
eng_best
equ
osd
```

Related to:

- https://github.com/scientist-softserv/adventist-dl/issues/248