Narrowing list of derivatives
Prior to this commit, we were generating several derivatives that are
likely unnecessary.  That, as [noted in Slack conversations][1], may
result in slower processing times.

With this change, we're favoring only the text extraction and instead
relying on the original files to be an adequate format.

This aligns with the proposed default configuration in:

- https://github.com/scientist-softserv/iiif_print/pull/181

[1]: https://assaydepot.slack.com/archives/C0313NJV9PE/p1678730644979469?thread_ts=1678729576.869239&cid=C0313NJV9PE