Updating Hyrax::DerivativeService.services spec
This is a provisional update; but one that I think requires
conversation.  Namely, should the `Hyrax::DerivativeService.services`
have anything besides `IiifPrint::PluggableDerivativeService`?

As implemented in Hyrax, the first service we encounter that is valid is
the one that is run.

The `Adventist::TextFileTextExtractionService` only runs against plain
text files.  However, I believe the
`IiifPrint::PluggableDerivativeService` wraps the
`Hyrax::DerivativeService`.

Thoughts?

Related to:

- https://github.com/scientist-softserv/adventist-dl/issues/317