Update LineageService and config
This commit will add an over to IiifPrint::LineageService to look for
slugs over ids.  Also, we found that in some cases using `--psm 1` had
problems where it flips the image incorrectly when the image is already
in the right orientation.  Considerations will be needed for images that
are actually in the wrong orientation though.