Update IIIF Manifest labels to show titles
Prior to this commit, the IIIF Manifest's canvas label is based on the
record's label which is actually the filename.  This is not desirable
because the filename is often times not human-readable.  This commit
will instead instead the file set's parent work's title and use that
instead as the label.