Adding logging for muted exception
Prior to this commit, we were taking an exception and passing just the
message to the user.  We abandoned the useful backtrace and context.

With this commit, we add error logging of the back trace.

An outstanding question is "What would be the consequence of re-raising
the exception?"  Would this help us gain better insight into what's
going on?  It seems like we probably should do that.