🎁 Adding offset to OAI parsing
With this commit, I'm adding the ability to specify an offset for the
OAI feed.  By default that offset is 0.

The ingest will skip over "offset" number of records before it starts
processing records.

To test this you will need to create an importer using the OAI parser,
with the following:

- URL :: http://oai.adventistdigitallibrary.org/OAI-script
- metadata_prefix: oai_adl

You'll need to run two tests:

- Limit 2, Offset 0 :: Will ingest the first two records
- Limit 1, Offset 1 :: Will attempt to ingest the second record only

You will want to then check the number of entries in the processed list;
not the count (which can be incorrect).

Related to:

- https://github.com/scientist-softserv/adventist-dl/issues/401