🎁 Repeatable service to re-run failed entries
During the normal lifecycle of an import, we inevitably encounter
errors.  Those errors can be from a variety of problems.

This service class, which is not part of production, is the beginnings
of the logic necessary for re-running only the failed entries while also
emulating having re-run all of the successful entries.

Part of the reason for the emulation is to ensure that counts, which are
already suspect, are consistent from run to run.

Note: I have prefaced this PR with the 🚧; to indicate "under
construction".  This is something that can be merged as it is not called
in production code.

Related to:

- https://github.com/scientist-softserv/adventist-dl/issues/430
- https://github.com/scientist-softserv/adventist-dl/issues/437