Remove rescue
Remove rescue and error handling from the job.
- Capturing and attempting to handle potential errors seemed to cause
unreliable results (collections were randomly being lost from some
subsequently processed works, though it was not a totally natural test,
since the errors were forced).
- In addition, it triggered the job to rerun 5 times,
but end with NO error.

By not capturing an error, we should end up with a job showing the
error and be able to rerun.