🐛 OCR count correction
This commit will bring in a fix from IIIF Print that will compensate for
the count being off due to using snippets.  The OCR text is indexed on
the FileSets of each work typically, but in order to get the snippets
working, all of the OCR need to also be indexed onto the parent work.
This will double the count effectively of everything because it's found
on the parent work and also on one of the FileSets.

Also in this commit, it was noted that the parent work started to index
the JSON coordinates and not the text itself.  This will be fixed with
this update as well.

Lastly, there was a change in IIIF Print that enabled host applications
to not display the snippet field if there are no snippets.  This will
bring that in as well as leverage that functionality.

Ref:
  - https://github.com/scientist-softserv/adventist-dl/issues/477